# [status-] decrease latency of status() when other threads run