--- a/visidata/features/status_source.py
+++ b/visidata/features/status_source.py
@@ -5,7 +5,7 @@ from visidata import VisiData
 
 @VisiData.api
 def getStatusSource(vd):
-    stack = inspect.stack()
+    stack = inspect.stack(context=0)  #2370
     for i, sf in enumerate(stack):
         if sf.function in 'status aside'.split():
             if stack[i+1].function in 'error fail warning debug'.split():