@saulpw 2024-07-10
I had some problems with `inspect.stack()` myself, in which a status message from `Ctrl+C` while loading hung the process indefinitely.  So I decided to only do it when `options.debug` is set.